requirements.txt
requests==2.28.1
python-dotenv==0.21.
python-dateutil==2.8.2
python-telegram-bot==13.12
